+++ src/main/java/com/openclassrooms/repository/MessageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.openclassrooms.model.Message;
 
 @Repository
-public interface MessageRepository extends CrudRepository<Message, Long> {
+public interface MessageRepository extends CrudRepository<Message, Integer> {
 
 }
