+++ src/main/java/com/openclassrooms/repository/RentalRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.openclassrooms.model.Rental;
 
 @Repository
-public interface RentalRepository extends CrudRepository<Rental, Long> {
+public interface RentalRepository extends CrudRepository<Rental, Integer> {
 
 }
